+++ lib/himg.rb
@@ -16,6 +16,7 @@ end
 #
 # Converts HTML to an Image for a minimal subset of HTML and CSS
 module Himg
+  RENDER_OPTIONS = %i[width height truncate verbose].freeze
   class Error < StandardError; end
 
   def self.render(html, width: 720, height: 405, truncate: true, verbose: false)