+++ lib/himg/railtie/template_handler.rb
@@ -4,7 +4,7 @@ module Himg
     class TemplateHandler
       def self.call(_template, source)
         <<-CODE
-          Himg.render(#{source.inspect})
+          Himg.render(#{source.inspect}, **@_himg_config)
         CODE
       end
     end
@@ -21,7 +21,7 @@ module Himg
           output = begin
             #{preprocessed_view_code}
           end
-          Himg.render(output)
+          Himg.render(output, **@_himg_config)
         CODE
       end
     end